+++ src/orego/mcts/RaveNodeTest.java
@@ -51,7 +51,7 @@ public class RaveNodeTest {
 	
 	@Test
 	public void testRecordTiedPlayout() {
-		node.recordPlayout(0.5, new int[] {at("a1"), at("b2"), at("a2")}, 0, 3, new IntSet(FIRST_POINT_BEYOND_BOARD));
+		node.recordPlayout((float) 0.5, new int[] {at("a1"), at("b2"), at("a2")}, 0, 3, new IntSet(FIRST_POINT_BEYOND_BOARD));
 		assertEquals(3, node.getRaveRuns(at("a1")), 0.001);
 		assertEquals(1.5, node.getRaveWins(at("a1")), 0.001);
 		assertEquals(3, node.getRaveRuns(at("a2")), 0.001);
